Control.Foldl.Except

Documentation

joinEither :: Either a a -> a

newtype FoldE a m b

Monadic fold with early result (before folding over an entire input)

For example, could define:

 countIndent :: FoldE Char m Int
 countIndent = FoldE (FoldM (i y -> if isSpace y then return (i+1) else throwE i) 0 return)

Constructors

FoldE

Fields

runFoldE :: FoldM (ExceptT b m) a b

extractE :: Monad m => FoldE a m b -> m b

Extract the current result

pureFoldE :: Monad m => b -> FoldE a m b

Consume no input and return the given value

returnFoldE :: Monad m => m b -> FoldE a m b

returnFoldE and pureFoldE should be pretty quick, they end the FoldE at every point.

foldEs :: (Monad m, Foldable t) => FoldE c m (Maybe b) -> t (FoldE c m (Maybe b)) -> CofreeF (MaybeT m) c -> CofreeF (MaybeT m) (Either b c)

Given a default and Foldable collection of ways to parse a b, try each (only rewinding up to a single char) and return (Left the_first_to_throwE (Just x)), or recurse.

foldEs_ :: (Monad m, Foldable t) => t (FoldE c m (Maybe b)) -> CofreeF (MaybeT m) c -> CofreeF (MaybeT m) (Either b c)

foldE1 :: Monad m => FoldE a m (Maybe c) -> CofreeF (MaybeT m) (Either b a) -> CofreeF (MaybeT m) (Either (Either b c) a)

foldE1 may be combined with foldE0 as follows to combine arbitrarily large numbers of FoldEs:

f g -> foldE1 g . foldE0 f
  :: Monad m =>
        FoldE a m (Maybe b)
     -> FoldE a m (Maybe c)
     -> CofreeF (MaybeT m) a
     -> CofreeF (MaybeT m) (Either (Either b c) a)

f g h -> foldE1 h . foldE1 g . foldE0 f
  :: Monad m =>
        FoldE a m (Maybe b)
     -> FoldE a m (Maybe c1)
     -> FoldE a m (Maybe c2)
     -> CofreeF (MaybeT m) a
     -> CofreeF (MaybeT m) (Either (Either (Either b c1) c2) a)

Or, we may not want to preserve distinct FoldE types:

f g -> fmap (first joinEither) . foldE1 g . foldE0 f
  :: Monad m =>
        FoldE c m (Maybe b)
     -> FoldE c m (Maybe b)
     -> CofreeF (MaybeT m) c
     -> CofreeF (MaybeT m) (Either b c)

f g h -> fmap (first joinEither) . foldE1 h . fmap (first joinEither) . foldE1 g . foldE0 f
  :: Monad m =>
        FoldE c m (Maybe b)
     -> FoldE c m (Maybe b)
     -> FoldE c m (Maybe b)
     -> CofreeF (MaybeT m) c
     -> CofreeF (MaybeT m) (Either b c)

foldE1' :: Monad m => (x -> a -> ExceptT (Maybe c) m x) -> x -> (x -> ExceptT (Maybe c) m (Maybe c)) -> Maybe (Cofree (MaybeT m) (Either b a)) -> m (Maybe (Cofree (MaybeT m) (Either (Either b c) a)))

foldE0 :: Monad m => FoldE a m (Maybe b) -> CofreeF (MaybeT m) a -> CofreeF (MaybeT m) (Either b a)

foldE0' :: Monad m => (x -> a -> ExceptT (Maybe b) m x) -> ExceptT (Maybe b) m x -> (x -> ExceptT (Maybe b) m (Maybe b)) -> Maybe (Cofree (MaybeT m) a) -> m (Maybe (Cofree (MaybeT m) (Either b a)))

newtype Fold1E a m b

FoldE for non-empty inputs

Constructors

Fold1E

Fields

runFold1E :: FoldM1 (ExceptT b m) a b

fold1E :: Monad m => b -> Fold1E a m b -> FoldE a m b

foldIf1 :: Monad m => (a -> Bool) -> Fold1E a m (Maybe b) -> FoldE a m (Maybe b)

span1E :: Monad m => (a -> Bool) -> (a -> a -> MaybeT m b) -> Fold1E a m (Maybe b)

span2E :: Monad m => (a -> a -> Bool) -> (a -> a -> MaybeT m b) -> Fold1E a m (Maybe b)

spanIfE2 :: Monad m => (a -> Bool) -> (a -> a -> Bool) -> (a -> a -> MaybeT m b) -> FoldE a m (Maybe b)

parseStringWith :: (Comonad w, Monad m) => (w Char -> w Char -> MaybeT m b) -> FoldE (w Char) m (Maybe b)

foldIf :: Monad m => (a -> Bool) -> FoldE a m b -> FoldE a m b

If empty, extractM, if the predicate passes then fold with the given fold else throwE.

takeWhile1 :: Monad m => (a -> Bool) -> FoldE a m b -> FoldE a m b

takeUntil1 :: Monad m => (a -> Bool) -> FoldE a m b -> FoldE a m b