Control.Foldl1

Documentation

newtype Fold1 a b

A Fold over non-empty inputs

Constructors

Fold1

Fields

runFold1 :: a -> Fold (a, a) b

newtype FoldM1 m a b

A FoldM over non-empty inputs

Constructors

FoldM1

Fields

runFoldM1 :: a -> FoldM m (a, a) b

toFold1 :: Fold a b -> Fold1 a b

toFoldM1 :: Monad m => FoldM m a b -> FoldM1 m a b

fold1 :: b -> Fold1 a b -> Fold a b

Given a default result, we can transform a Fold that requires an initial value and folds over (prev value, current value) into a simple Fold.

foldM1 :: Monad m => b -> FoldM1 m a b -> FoldM m a b

foldPairF for FoldM